requirements.txt
selenium==3.4.1
lxml==3.7.2
selenium_requests==1.3
requests==2.12.4
discord.py==0.16.7
google_api_python_client==1.6.2
